+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [1.0.7] 2023-10-30
 
+- Improve Chart tooltips
 - Migrate to sqlite3 & sqlite3_flutter_libs
 
 ## [1.0.6] 2023-10-29